requirements.txt
# Core
numpy<2.0
pandas==2.2.3
tqdm==4.67.1
tabulate==0.9.0

# Dataset (BEIR)
beir==1.0.1

# Retrieval backends (implemented later)
elasticsearch>=7.9.1
sentence-transformers==3.2.1
faiss-cpu>=1.8.0

# Dev / tests
pytest==8.3.4
matplotlib>=3.8.0
